+++ packages/test-renderer/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @react-three/test-renderer
 
+## 7.0.12
+
+### Patch Changes
+
+- 0df6073: fix: missed events
+- Updated dependencies [0df6073]
+  - @react-three/fiber@7.0.12
+
+## 7.0.11
+
+### Patch Changes
+
+- 62b0a3a: fix: event order of missed pointers
+- Updated dependencies [62b0a3a]
+  - @react-three/fiber@7.0.11
+
+## 7.0.10
+
+### Patch Changes
+
+- e019dd4: fixes
+- Updated dependencies [e019dd4]
+  - @react-three/fiber@7.0.10
+
+## 7.0.8
+
+### Patch Changes
+
+- cd266e4: Fix diffProps dashed keys
+- Updated dependencies [cd266e4]
+  - @react-three/fiber@7.0.9
+
 ## 7.0.7
 
 ### Patch Changes